TAT/TaskAnalyzer/requirements.txt
anthropic==0.9.0
Pillow==10.2.0
